services/likes_service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9
channels>=4.1
channels-redis>=4.2
uvicorn[standard]>=0.30
djangorestframework-simplejwt>=5.3
redis>=5.0
django-cors-headers

